+++ R/analysis_fit_methods_BASE_1240.R
@@ -253,12 +253,16 @@ fit_method.wand <- function(dat, ...) {
 
   dots <- list(...)
 
-  if(is.null(dots$pop_k))
-    dots$pop_k <- 10
-  if(is.null(dots$subj_k))
-    dots$subj_k <- 5
+  pop_k <- 10
+  subj_k <- 5
 
-  fit_obj <- do.call(wand_fit, c(as.list(dat), dots))
+  if(!is.null(dots$pop_k))
+    pop_k <- 2
+  if(!is.null(dots$subj_k))
+    subj_k <- 2
+
+  fit_obj <- wand_fit(dat$x, dat$y, dat$subjid,
+    pop_k = pop_k, subj_k = subj_k, ...)
 
   fit_apply <- function(dat, xg = NULL, cpx = NULL, fit) {