+++ R/plot_trajectory.R
@@ -8,17 +8,26 @@
 #' @param hover variable names in \code{x$data} to show on hover for each point (only variables with non-NA data will be shown)
 #' @param checkpoints should the checkpoints be plotted (if available)?
 #' @param p centiles at which to draw the WHO polygons
+#' @param x_units units of age x-axis (days, months, or years)
 #' @param \ldots additional parameters passed to \code{\link{figure}}
 #' @examples
 #' mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
 #' fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
 #' plot(fit)
+#' plot(fit, x_units = "years")
 #' plot(fit, center = TRUE)
 #' plot(fit, hover = c("wtkg", "bmi", "waz", "haz"))
 #' @export
 plot.fittedTrajectory <- function(x, center = FALSE, x_range = NULL,
   width = 500, height = 520, hover = NULL, checkpoints = TRUE,
-  p = 100 * pnorm(-3:0), ...) {
+  p = 100 * pnorm(-3:0),
+  x_units = c("days", "months", "years"), ...) {
+
+  x_units <- match.arg(x_units)
+  x_denom <- switch(x_units,
+    days = 1,
+    months = 365.25 / 12,
+    years = 365.25)
 
   if(nrow(x$xy) == 0)
     return(empty_plot(paste0("No '", x$y_var, "' vs. '", x$x_var, "' data for this subject")))
@@ -32,14 +41,15 @@ plot.fittedTrajectory <- function(x, center = FALSE, x_range = NULL,
   #   hover <- names(x$data)[sapply(x$data, function(x) !all(is.na(x)))]
   #   hover <- x$data[x$xy$idx, hover]
   # } else
-  if(!is.null(hover)) {
-    hover <- intersect(names(x$data), hover)
-    if(length(hover) == 0) {
-      hover <- NULL
-    } else {
-      hover <- hover[sapply(x$data[,hover], function(x) !all(is.na(x)))]
-      hover <- x$data[x$xy$idx, hover]
-    }
+  if(is.null(hover))
+    hover <- c(x$x_var, x$y_var)
+
+  hover <- intersect(names(x$data), hover)
+  if(length(hover) == 0) {
+    hover <- NULL
+  } else {
+    hover <- hover[sapply(x$data[,hover], function(x) !all(is.na(x)))]
+    hover <- x$data[x$xy$idx, hover]
   }
 
   ylab <- hbgd::hbgd_labels[[x$y_var]]
@@ -57,19 +67,25 @@ plot.fittedTrajectory <- function(x, center = FALSE, x_range = NULL,
     ylab <- paste(ylab, "(WHO median-centered)")
   }
 
+  xlab <- hbgd::hbgd_labels[[x$x_var]]
+  if(x_units == "months")
+    xlab <- gsub("\\(days\\)", "(months)", xlab)
+  if(x_units == "years")
+    xlab <- gsub("\\(days\\)", "(years)", xlab)
+
   fig <- figure(width = width, height = height,
-    xlab = hbgd::hbgd_labels[[x$x_var]], ylab = ylab, logo = NULL, ...) %>%
+    xlab = xlab, ylab = ylab, logo = NULL, ...) %>%
     ly_who(x = seq(x_range[1], x_range[2], length = 100), center = center,
-      x_var = x$x_var, y_var = x$y_var, sex = x$sex, p = p) %>%
-    rbokeh::ly_points(x, y, hover = hover, data = x$xy, color = "black")
+      x_var = x$x_var, y_var = x$y_var, sex = x$sex, p = p, x_units = x_units) %>%
+    rbokeh::ly_points(x / x_denom, y, hover = hover, data = x$xy, color = "black")
   if(!is.null(x$fitgrid)) {
     fig <- fig %>%
-      rbokeh::ly_lines(x, y, data = x$fitgrid, color = "black") %>%
-      rbokeh::ly_points(x, yfit, data = x$xy, color = "black", glyph = 19, size = 4)
+      rbokeh::ly_lines(x / x_denom, y, data = x$fitgrid, color = "black") %>%
+      rbokeh::ly_points(x / x_denom, yfit, data = x$xy, color = "black", glyph = 19, size = 4)
   }
   if(!is.null(x$holdout))
     fig <- fig %>%
-      rbokeh::ly_points(x, y, data = x$holdout, color = "red")
+      rbokeh::ly_points(x / x_denom, y, data = x$holdout, color = "red")
 
   if(!all(is.na(x$checkpoint$y)) && checkpoints) {
     x$checkpoint <- subset(x$checkpoint, !is.na(y))
@@ -77,7 +93,7 @@ plot.fittedTrajectory <- function(x, center = FALSE, x_range = NULL,
     x$checkpoint$zcat <- as.character(x$checkpoint$zcat)
 
     fig <- fig %>%
-      rbokeh::ly_points(x, y, size = 15, hover = c("zcat", "x"),
+      rbokeh::ly_points(x / x_denom, y, size = 15, hover = c("zcat", "x"),
         data = x$checkpoint, glyph = 13, color = "black", alpha = 0.6)
   }
 
@@ -89,19 +105,30 @@ plot.fittedTrajectory <- function(x, center = FALSE, x_range = NULL,
 #' @param x an object returned from \code{\link{fit_trajectory}}
 #' @param x_range a vector specifying the range (min, max) that the superposed z-score bands should span on the x-axis
 #' @param nadir should a guide be added to the plot showing the location of the nadir?
+#' @param recovery age in days at which to plot recovery from nadir (only valid if nadir is TRUE) - if NULL (default), will not be plotted
 #' @param width width of the plot
 #' @param height height of the plot
 #' @param hover variable names in \code{x$data} to show on hover for each point (only variables with non-NA data will be shown)
 #' @param checkpoints should the checkpoints be plotted (if available)?
 #' @param z z-scores at which to draw the z-score bands
+#' @param x_units units of age x-axis (days, months, or years)
 #' @param \ldots additional parameters passed to \code{\link{figure}}
 #' @examples
 #' mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
 #' fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
 #' plot_z(fit)
 #' @export
-plot_z <- function(x, x_range = NULL, nadir = FALSE, width = 500, height = 520,
-  hover = NULL, checkpoints = TRUE, z = -3:0, ...) {
+plot_z <- function(x, x_range = NULL, nadir = FALSE, recovery = NULL,
+  width = 500, height = 520,
+  hover = NULL, checkpoints = TRUE, z = -3:0,
+  x_units = c("days", "months", "years"), ...) {
+
+  x_units <- match.arg(x_units)
+  x_denom <- switch(x_units,
+    days = 1,
+    months = 365.25 / 12,
+    years = 365.25)
+
   if(is.null(x$xy$z))
     return(empty_plot("No z transformation data for this subject"))
 
@@ -110,45 +137,68 @@ plot_z <- function(x, x_range = NULL, nadir = FALSE, width = 500, height = 520,
     x_range <- x_range + c(-1, 1) * diff(x_range) * 0.07
   }
 
-  if(!is.null(hover)) {
-    hover <- intersect(names(x$data), hover)
-    if(length(hover) == 0) {
-      hover <- NULL
-    } else {
-      hover <- hover[sapply(x$data[,hover], function(x) !all(is.na(x)))]
-      hover <- x$data[x$xy$idx, hover]
-    }
+  if(is.null(hover)) {
+    y_var_out <- x$y_var
+    if(x$y_var == "htcm")
+      y_var_out <- "haz"
+    if(x$y_var == "wtkg")
+      y_var_out <- "waz"
+    hover <- c(x$x_var, y_var_out)
+  }
+  hover <- intersect(names(x$data), hover)
+  if(length(hover) == 0) {
+    hover <- NULL
+  } else {
+    hover <- hover[sapply(x$data[,hover], function(x) !all(is.na(x)))]
+    hover <- x$data[x$xy$idx, hover]
   }
 
   xlab <- hbgd::hbgd_labels[[x$x_var]]
+  if(x_units == "months")
+    xlab <- gsub("\\(days\\)", "(months)", xlab)
+  if(x_units == "years")
+    xlab <- gsub("\\(days\\)", "(years)", xlab)
   ylab <- paste(hbgd::hbgd_labels[[x$y_var]], "z-score")
 
   fig <- figure(width = width, height = height,
     xlab = xlab, ylab = ylab, logo = NULL, ...) %>%
     ly_zband(x = c(x_range[1], x_range[2]), z = z,
-      color = ifelse(x$sex == "Male", "blue", "red")) %>%
-    rbokeh::ly_points(x, z, hover = hover, data = x$xy, color = "black")
+      color = ifelse(x$sex == "Male", "blue", "red"), x_units = x_units) %>%
+    rbokeh::ly_points(x / x_denom, z, hover = hover, data = x$xy, color = "black")
   if(!is.null(x$fitgrid)) {
     fig <- fig %>%
-      rbokeh::ly_lines(x, z, data = x$fitgrid, color = "black") %>%
-      rbokeh::ly_points(x, zfit, data = x$xy, color = "black", glyph = 19, size = 4)
+      rbokeh::ly_lines(x / x_denom, z, data = x$fitgrid, color = "black") %>%
+      rbokeh::ly_points(x / x_denom, zfit, data = x$xy, color = "black",
+        glyph = 19, size = 4)
   }
   if(!is.null(x$holdout))
     fig <- fig %>%
-      rbokeh::ly_points(x, z, data = x$holdout, color = "red")
+      rbokeh::ly_points(x / x_denom, z, data = x$holdout, color = "red")
 
   if(!all(is.na(x$checkpoint$y)) && checkpoints) {
     x$checkpoint <- subset(x$checkpoint, !is.na(y))
     fig <- fig %>%
-      rbokeh::ly_points(x, z, size = 15, hover = zcat, data = x$checkpoint, glyph = 13, color = "black", alpha = 0.6)
+      rbokeh::ly_points(x / x_denom, z, size = 15, hover = zcat, data = x$checkpoint, glyph = 13, color = "black", alpha = 0.6)
   }
 
   if(nadir) {
     nadir <- get_nadir(x)
     if(!is.na(nadir$at)) {
       fig <- fig %>%
-        rbokeh::ly_segments(nadir$at, 0, nadir$at, nadir$mag, line_width = 4,
-          color = "black", alpha = 0.3)
+        rbokeh::ly_segments(nadir$at / x_denom, 0, nadir$at / x_denom, nadir$mag, line_width = 5, color = "red", alpha = 0.5)
+
+      if(!is.null(recovery)) {
+        recov <- get_recovery(x, nadir, recovery)
+        if(!is.na(recov$at)) {
+          fig <- fig %>%
+            rbokeh::ly_segments(nadir$at / x_denom, nadir$mag,
+              recov$at / x_denom, nadir$mag,
+              width = 5, color = "orange", alpha = 0.5) %>%
+            rbokeh::ly_segments(recov$at / x_denom, nadir$mag,
+              recov$at / x_denom, recov$z,
+              width = 5, color = "green", alpha = 0.5)
+        }
+      }
     }
   }
 
@@ -160,17 +210,30 @@ plot_z <- function(x, x_range = NULL, nadir = FALSE, width = 500, height = 520,
 #' @param x an object returned from \code{\link{fit_trajectory}}
 #' @param width width of the plot
 #' @param height height of the plot
+#' @param x_units units of age x-axis (days, months, or years)
 #' @param \ldots additional parameters passed to \code{\link{figure}}
 #' @examples
 #' mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
 #' fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
 #' plot_velocity(fit)
 #' @export
-plot_velocity <- function(x, width = 500, height = 520, ...) {
+plot_velocity <- function(x, width = 500, height = 520,
+  x_units = c("days", "months", "years"), ...) {
+
+  x_units <- match.arg(x_units)
+  x_denom <- switch(x_units,
+    days = 1,
+    months = 365.25 / 12,
+    years = 365.25)
+
   if(is.null(x$fitgrid$dy))
     return(empty_plot("No velocity data for this subject"))
 
   xlab <- hbgd::hbgd_labels[[x$x_var]]
+  if(x_units == "months")
+    xlab <- gsub("\\(days\\)", "(months)", xlab)
+  if(x_units == "years")
+    xlab <- gsub("\\(days\\)", "(years)", xlab)
   ylab <- paste(hbgd::hbgd_labels[[x$y_var]], "growth velocity")
 
   # remove blip in velocity
@@ -183,7 +246,7 @@ plot_velocity <- function(x, width = 500, height = 520, ...) {
 
   figure(width = width, height = height,
     xlab = xlab, ylab = ylab, logo = NULL, ...) %>%
-    ly_lines(xx, dyy, color = "black")
+    ly_lines(xx / x_denom, dyy, color = "black")
 }
 
 #' Plot a fitted trajectory's z-score velocity
@@ -191,17 +254,30 @@ plot_velocity <- function(x, width = 500, height = 520, ...) {
 #' @param x an object returned from \code{\link{fit_trajectory}}
 #' @param width width of the plot
 #' @param height height of the plot
+#' @param x_units units of age x-axis (days, months, or years)
 #' @param \ldots additional parameters passed to \code{\link{figure}}
 #' @examples
 #' mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
 #' fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
 #' plot_zvelocity(fit)
 #' @export
-plot_zvelocity <- function(x, width = 500, height = 520, ...) {
+plot_zvelocity <- function(x, width = 500, height = 520,
+  x_units = c("days", "months", "years"), ...) {
+
+  x_units <- match.arg(x_units)
+  x_denom <- switch(x_units,
+    days = 1,
+    months = 365.25 / 12,
+    years = 365.25)
+
   if(is.null(x$fitgrid$dz))
     return(empty_plot("No z-score velocity data for this subject"))
 
   xlab <- hbgd::hbgd_labels[[x$x_var]]
+  if(x_units == "months")
+    xlab <- gsub("\\(days\\)", "(months)", xlab)
+  if(x_units == "years")
+    xlab <- gsub("\\(days\\)", "(years)", xlab)
   ylab <- paste(hbgd::hbgd_labels[[x$y_var]], "z-score growth velocity")
 
   # remove blip in velocity
@@ -214,7 +290,7 @@ plot_zvelocity <- function(x, width = 500, height = 520, ...) {
 
   figure(width = width, height = height,
     xlab = xlab, ylab = ylab, logo = NULL, ...) %>%
-    ly_lines(xx, dzz, color = "black")
+    ly_lines(xx / x_denom, dzz, color = "black")
 }
 
 empty_plot <- function(lab) {
@@ -223,21 +299,60 @@ empty_plot <- function(lab) {
     ly_text(0, 0, c("", lab), align = "center")
 }
 
-#' Get nadir of a growth velocity
+#' Get nadir of z-scale growth trajectory
 #'
 #' @param obj object created from \code{\link{fit_trajectory}}
 #' @export
 get_nadir <- function(obj) {
   if(is.null(obj$fitgrid))
-    return(list(at = NA, mag = NA))
+    return(data.frame(at = NA, mag = NA, end = NA))
   if(is.null(obj$fitgrid$dz))
-    return(list(at = NA, mag = NA))
+    return(data.frame(at = NA, mag = NA, end = NA))
+
+  nn <- nrow(obj$fitgrid) - 1
 
   # get crossings of zero of dz
-  cross <- which(diff(sign(obj$fitgrid$dz)) != 0)
-  if(length(cross) == 0)
-    return(list(at = NA, mag = NA))
+  cross <- which(diff(sign(obj$fitgrid$dz)) > 0) + 1
+  if(length(cross) == 0) {
+    if(all(obj$fitgrid$dz[nn] >= obj$fitgrid$dz, na.rm = TRUE)) {
+      return(data.frame(at = obj$fitgrid$x[nn], mag = obj$fitgrid$z[nn], end = TRUE))
+    } else {
+      return(data.frame(at = NA, mag = NA, end = NA))
+    }
+  }
 
-  cross <- cross[1]
-  list(at = obj$fitgrid$x[cross], mag = obj$fitgrid$z[cross])
+  cross <- cross[which.min(obj$fitgrid$z[cross])]
+  end <- FALSE
+  if(obj$fitgrid$z[nn] < obj$fitgrid$z[cross]) {
+    cross <- nn
+    end <- TRUE
+  }
+
+  data.frame(at = obj$fitgrid$x[cross], mag = obj$fitgrid$z[cross], end = end)
 }
+
+#' Get recovery statistics of z-scale growth trajectory
+#'
+#' @param obj object created from \code{\link{fit_trajectory}}
+#' @param nadir object created from \code{\link{get_nadir}} (if NULL, will be automatically generated)
+#' @param at age (in days) at which to estimate recovery
+#' @export
+get_recovery <- function(obj, nadir = NULL, at = 365.25 * 3) {
+
+  if(is.null(obj$fitgrid))
+    return(data.frame(at = NA, mag = NA, end = FALSE))
+  if(is.null(obj$fitgrid$z))
+    return(data.frame(at = NA, mag = NA, end = FALSE))
+
+  if(is.null(nadir)) {
+    nadir <- get_nadir(obj)
+  }
+
+  if(!is.na(nadir$at) && nadir$at < at) {
+    val <- approxfun(obj$fitgrid$x, obj$fitgrid$z)(at)
+    return(data.frame(at = at, z = val, recov = val - nadir$mag))
+  } else {
+    return(data.frame(at = NA, z = NA, recov = NA))
+  }
+}
+